Move test_helpers into new test_utils module
