+++ aquadoggo/src/db/stores/test_utils/runner.rs
@@ -17,7 +17,7 @@ use crate::db::stores::test_utils::{TestData, TestDatabase};
 use crate::db::Pool;
 use crate::db::SqlStore;
 use crate::schema::SchemaProvider;
-use crate::test_helpers::{initialize_db, initialize_db_with_url};
+use crate::test_utils::{initialize_db, initialize_db_with_url};
 
 use super::{doggo_fields, doggo_schema};
