+++ aquadoggo/src/graphql/client/dynamic_query.rs
@@ -428,7 +428,7 @@ mod test {
     use crate::db::stores::test_utils::{
         add_document, add_schema, test_db, TestDatabase, TestDatabaseRunner,
     };
-    use crate::test_helpers::graphql_test_client;
+    use crate::test_utils::graphql_test_client;
 
     #[rstest]
     // Note: This and more tests in this file use the underlying static schema provider which is a
