+++ aquadoggo/src/graphql/client/tests.rs
@@ -14,7 +14,7 @@ use serial_test::serial;
 use crate::db::stores::test_utils::{
     add_document, add_schema, test_db, TestDatabase, TestDatabaseRunner,
 };
-use crate::test_helpers::graphql_test_client;
+use crate::test_utils::graphql_test_client;
 
 // Test querying application documents with scalar fields (no relations) by document id and by view
 // id.
