+++ aquadoggo/src/graphql/schema.rs
@@ -205,7 +205,7 @@ mod test {
     use serial_test::serial;
 
     use crate::db::stores::test_utils::{add_schema, test_db, TestDatabase, TestDatabaseRunner};
-    use crate::test_helpers::graphql_test_client;
+    use crate::test_utils::graphql_test_client;
 
     #[rstest]
     // Note: This test uses the underlying static schema provider which is a static mutable data
