+++ aquadoggo/src/http/service.rs
@@ -83,7 +83,7 @@ mod tests {
     use crate::graphql::GraphQLSchemaManager;
     use crate::http::context::HttpServiceContext;
     use crate::schema::SchemaProvider;
-    use crate::test_helpers::TestClient;
+    use crate::test_utils::TestClient;
 
     use super::build_server;
