+++ aquadoggo/src/lib.rs
@@ -57,8 +57,6 @@ mod node;
 mod replication;
 mod schema;
 #[cfg(test)]
-mod test_helpers;
-#[cfg(test)]
 mod test_utils;
 #[cfg(test)]
 mod tests;
